# Conflicts:
#	src/main/java/org/example/hospital_admission_project/repo/ChatRepository.java
#	src/main/java/org/example/hospital_admission_project/service/ChatService.java
#	src/main/java/org/example/hospital_admission_project/service/impl/AuthServiceImpl.java

+++ src/main/java/org/example/hospital_admission_project/controller/AppointmentController.java
@@ -14,7 +14,7 @@ public class AppointmentController {
     private final AppointmentService appointmentService;
 
     @PostMapping("/booking")
-    public ResponseEntity<?> createBooking(@RequestParam Integer doctorId, @RequestParam String reason) {
+    public ResponseEntity<?> createBooking(@RequestParam Long doctorId, @RequestParam String reason) {
         return appointmentService.createBooking(doctorId, reason);
     }
